Feature/room

+++ src/main/resources/db/migration/V1__init_schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE users
 (
-    id         SERIAL PRIMARY KEY,
+    id         BIGSERIAL PRIMARY KEY,
     first_name VARCHAR(50)  NOT NULL,
     last_name  VARCHAR(50)  NOT NULL,
     email      VARCHAR(150) NOT NULL UNIQUE,
@@ -12,7 +12,7 @@ CREATE TABLE users
 
 CREATE TABLE rooms
 (
-    id          SERIAL PRIMARY KEY,
+    id          BIGSERIAL PRIMARY KEY,
     name        VARCHAR(100) NOT NULL,
     description TEXT,
     capacity    INTEGER      NOT NULL CHECK (capacity >= 1 AND capacity <= 1000),
@@ -37,14 +37,14 @@ CREATE TABLE room_equipment
 
 CREATE TABLE reservations
 (
-    id                SERIAL PRIMARY KEY,
+    id                BIGSERIAL PRIMARY KEY,
     start_date_time   TIMESTAMP    NOT NULL,
     end_date_time     TIMESTAMP    NOT NULL,
     status            VARCHAR(20)  NOT NULL DEFAULT 'PENDING' CHECK (status IN ('PENDING', 'CONFIRMED', 'CANCELLED', 'REJECTED')),
     purpose           VARCHAR(255),
     number_of_people  INTEGER      NOT NULL,
-    user_id           INTEGER      NOT NULL REFERENCES users (id) ON DELETE RESTRICT,
-    room_id           INTEGER      NOT NULL REFERENCES rooms (id) ON DELETE RESTRICT,
+    user_id           BIGINT       NOT NULL REFERENCES users (id) ON DELETE RESTRICT,
+    room_id           BIGINT       NOT NULL REFERENCES rooms (id) ON DELETE RESTRICT,
     created_at        TIMESTAMP    NOT NULL DEFAULT NOW()
 );
 
@@ -53,12 +53,12 @@ CREATE INDEX idx_reservations_user_id ON reservations (user_id);
 
 CREATE TABLE files
 (
-    id           SERIAL PRIMARY KEY,
+    id           BIGSERIAL PRIMARY KEY,
     filename     VARCHAR(255) NOT NULL,
     stored_path  VARCHAR(500) NOT NULL,
     content_type VARCHAR(100) NOT NULL,
     size         BIGINT       NOT NULL,
-    room_id      INTEGER      NOT NULL REFERENCES rooms (id) ON DELETE CASCADE,
+    room_id      BIGINT       NOT NULL REFERENCES rooms (id) ON DELETE CASCADE,
     created_at   TIMESTAMP    NOT NULL DEFAULT NOW()
 );
